requirements.txt
aiohttp
vcrpy
